refactor: options into their own files.
