+++ packages/cspell/src/options.ts
@@ -1,0 +1,45 @@
+export interface CSpellApplicationOptions extends BaseOptions {
+    /**
+     * Display verbose information
+     */
+    verbose?: boolean;
+    /**
+     * Show extensive output.
+     */
+    debug?: boolean;
+    /**
+     * a globs to exclude files from being checked.
+     */
+    exclude?: string[] | string;
+    /**
+     * Only report the words, no line numbers or file names.
+     */
+    wordsOnly?: boolean;
+    /**
+     * unique errors per file only.
+     */
+    unique?: boolean;
+    /**
+     * root directory, defaults to `cwd`
+     */
+    root?: string;
+    /**
+     * Show part of a line where an issue is found.
+     * if true, it will show the default number of characters on either side.
+     * if a number, it will shat number of characters on either side.
+     */
+    showContext?: boolean | number;
+    /**
+     * Show suggestions for spelling errors.
+     */
+    showSuggestions?: boolean;
+}
+
+export type TraceOptions = BaseOptions;
+
+export interface BaseOptions {
+    config?: string;
+    languageId?: string;
+    locale?: string;
+    local?: string; // deprecated
+}
